# whatsudow_backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Whatsudow",
        "description": "Aplicación que permite a los organizadores de eventos unificar los datos de sus proveedores de servicios. Con esta app, los organizadores de eventos pueden almacenar y acceder a información importante de los proveedores, comocalendarios de disponibilidad, PDF actualizados, cláusulas y datos de contacto.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/v1/phones": {
            "get": {
                "tags": [
                    "Phones"
                ],
                "summary": "Obtener lista de teléfonos",
                "operationId": "aaab44c7e8dc61aa1ee1da3f7c557a37",
                "responses": {
                    "200": {
                        "description": "Éxito"
                    }
                }
            },
            "post": {
                "tags": [
                    "Phones"
                ],
                "summary": "Guardar teléfono",
                "operationId": "6385abccf161c54df6e6528cbe011565",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Éxito"
                    },
                    "422": {
                        "description": "Teléfono no válido"
                    }
                }
            }
        },
        "/api/v1/phones/{id}": {
            "get": {
                "tags": [
                    "Phones"
                ],
                "summary": "Obtener teléfono",
                "operationId": "5c0a6d0948f2071d9e2ed39420ac7b7e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del teléfono",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Éxito"
                    },
                    "404": {
                        "description": "No se encontró el teléfono del usuario"
                    }
                }
            },
            "put": {
                "tags": [
                    "Phones"
                ],
                "summary": "Actualizar teléfono",
                "operationId": "d646866aa7e6f5ce4f2444a7f18e4126",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del teléfono",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Éxito"
                    },
                    "404": {
                        "description": "No se encontró el teléfono del usuario"
                    },
                    "422": {
                        "description": "Teléfono no válido"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Phones"
                ],
                "summary": "Eliminar teléfono",
                "operationId": "e0eae1c2424dbab9e1879ae2b25802c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del teléfono",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Teléfono eliminado exitosamente"
                    },
                    "404": {
                        "description": "Teléfono no encontrado"
                    }
                }
            }
        },
        "/api/v1/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Obtener todos los servicios",
                "operationId": "19339d2b804c5bbb3a7d51826c6a5e9c",
                "responses": {
                    "200": {
                        "description": "Éxito"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Crear un nuevo servicio",
                "operationId": "b9fd7c793f65af72b2c4e91b4773220a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Éxito"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Datos de solicitud no válidos"
                    }
                }
            }
        },
        "/api/v1/services/{id}": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Obtener detalles de un servicio",
                "operationId": "5460efdf46e38dba4e94cc1c62ce15a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Éxito"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Services"
                ],
                "summary": "Actualizar un servicio",
                "operationId": "ca0d914039198efb0d6a0a5242999a96",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Éxito"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    },
                    "422": {
                        "description": "Datos de solicitud no válidos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Services"
                ],
                "summary": "Eliminar un servicio",
                "operationId": "284f2fc04534a4a5d77b970203f4e892",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Servicio eliminado exitosamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                }
            }
        },
...
    },
...
}